Defer creating the special shaders in the texture blitter
Won't apply as-is to Qt 6, needs a dedicated patch there. The urgency is
different there as well, from 6.4 on the blitter is not used at all in
the backingstore compositor, so the fact that preparing some of the
special shaders fail with some Intel drivers does not matter and
won't break applications using QOpenGLWidget or QQuickWidget.

Instead of introducing platform specific ifdefs (e.g. only supporting
texture rectangles on macOS, which is where it really is used), defer
the compilation of the more exotic shader variants.

Deferring the compilation of the potentially problematic
shaders to a point where they are actually needed has the benefit
of not affecting applications that never need those shaders to
begin with, i.e. even if the shader would fail to compile it won't
be a problem as the shader is never needed in the first place. In
addition, this provides a performance improvement to create() because
it now only compiles the commonly used shader, not the other two.

This way, even if the sampler2DRect-based shader failed
to compile, that would not be a problem for an application using
a QOpenGLWidget (and so using the blitter internally), because
that particular shader is not needed at all for rendering.

What exactly fails with certain older (but not that old) Intel
drivers, and what "syntax error syntax error" really means remain
a mystery, and likely not something we will solve.

Fixes: QTBUG-101396
Change-Id: Ic923d51007af503f2749eaae36297ca6647cfa09
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>