Doc: Add missing since 6.2 to QImage and QPixmap
The deviceIndependentSize() function of QImage and QPixmap was added
in Qt 6.2 but their docs are missing this information, so add them.

And it seems the QImage docs were copied from QPixmap, so it still
reference to pixmap, not image, fix it as a drive-by.

Change-Id: Ic3d0a2019b5e9b89bd8723d994d7ae54f2cb5257
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 0d20e6e4957b3842b659522f330b8a201ef8e505)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>