QOffsetStringArray: simplify makeOffsetStringArray()
Instead of building two arrays, a short one for size_t, and then a
full-length one for the final narrowed MinifiedOffsetType, construct
the final array directly.

This should also speed up compiling qOffsetStringArray() expressions,
Clang's -ftime-trace, for ninja tst_qoffsetstringarray, with a dirty
qoffsetstringarray_p.h (which isn't included in the PCH, so only
recompiles TUs that actually include it) shows a slight improvement:

 -   37 ms: qOffsetStringArray<$> (7 times, avg 5 ms)
 +   35 ms: qOffsetStringArray<$> (7 times, avg 5 ms)

I had to disable one static_assert because it looks to be running over
some GCC limitation and concluding the array is not constexpr (even
though it is).

qhttpheaders.cpp:486:29: error: non-constant condition for static assertion
qhttpheaders.cpp:486:29: error: the value of ‘headerNames’ is not usable in a constant expression

Change-Id: I0d6109645075e5b533391ea04ea845538ad496fe
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>